+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/InlineQueryTriggers.kt
@@ -3,6 +3,7 @@ package dev.inmo.tgbotapi.extensions.behaviour_builder.triggers_handling
 import dev.inmo.micro_utils.coroutines.subscribeSafelyWithoutExceptionsAsync
 import dev.inmo.tgbotapi.extensions.behaviour_builder.*
 import dev.inmo.tgbotapi.extensions.behaviour_builder.expectations.expectFlow
+import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.SimpleFilter
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.marker_factories.ByUserInlineQueryMarkerFactory
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.marker_factories.MarkerFactory
 import dev.inmo.tgbotapi.extensions.utils.asInlineQueryUpdate
@@ -11,7 +12,7 @@ import dev.inmo.tgbotapi.types.InlineQueries.query.*
 
 internal suspend inline fun <reified T : InlineQuery> BehaviourContext.onInlineQuery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
-    noinline additionalFilter: (suspend (T) -> Boolean)? = null,
+    noinline additionalFilter: SimpleFilter<T>? = null,
     markerFactory: MarkerFactory<in T, Any> = ByUserInlineQueryMarkerFactory,
     noinline scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, T>
 ) = flowsUpdatesFilter.expectFlow(bot) {
@@ -41,7 +42,7 @@ internal suspend inline fun <reified T : InlineQuery> BehaviourContext.onInlineQ
 
 suspend fun BehaviourContext.onAnyInlineQuery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
-    additionalFilter: (suspend (InlineQuery) -> Boolean)? = null,
+    additionalFilter: SimpleFilter<InlineQuery>? = null,
     markerFactory: MarkerFactory<in InlineQuery, Any> = ByUserInlineQueryMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, InlineQuery>
 ) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, markerFactory, scenarioReceiver)
@@ -49,7 +50,7 @@ suspend fun BehaviourContext.onAnyInlineQuery(
 
 suspend fun BehaviourContext.onBaseInlineQuery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
-    additionalFilter: (suspend (BaseInlineQuery) -> Boolean)? = null,
+    additionalFilter: SimpleFilter<BaseInlineQuery>? = null,
     markerFactory: MarkerFactory<in BaseInlineQuery, Any> = ByUserInlineQueryMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, BaseInlineQuery>
 ) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, markerFactory, scenarioReceiver)
@@ -57,7 +58,7 @@ suspend fun BehaviourContext.onBaseInlineQuery(
 
 suspend fun BehaviourContext.onLocationInlineQuery(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
-    additionalFilter: (suspend (LocationInlineQuery) -> Boolean)? = null,
+    additionalFilter: SimpleFilter<LocationInlineQuery>? = null,
     markerFactory: MarkerFactory<in LocationInlineQuery, Any> = ByUserInlineQueryMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, LocationInlineQuery>
 ) = onInlineQuery(includeFilterByChatInBehaviourSubContext, additionalFilter, markerFactory, scenarioReceiver)