+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/PassportTriggers.kt
@@ -3,6 +3,7 @@ package dev.inmo.tgbotapi.extensions.behaviour_builder.triggers_handling
 import dev.inmo.micro_utils.coroutines.subscribeSafelyWithoutExceptionsAsync
 import dev.inmo.tgbotapi.extensions.behaviour_builder.*
 import dev.inmo.tgbotapi.extensions.behaviour_builder.expectations.expectFlow
+import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.SimpleFilter
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.marker_factories.ByChatMessageMarkerFactory
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.marker_factories.MarkerFactory
 import dev.inmo.tgbotapi.extensions.utils.asMessageUpdate
@@ -13,7 +14,7 @@ import dev.inmo.tgbotapi.types.passport.encrypted.abstracts.EncryptedPassportEle
 
 internal suspend inline fun <reified T : EncryptedPassportElement> BehaviourContext.onPassportMessageWith(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
-    noinline additionalFilter: (suspend (PassportMessage) -> Boolean)? = null,
+    noinline additionalFilter: SimpleFilter<PassportMessage>? = null,
     markerFactory: MarkerFactory<in PassportMessage, Any> = ByChatMessageMarkerFactory,
     noinline scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, PassportMessage>
 ) = flowsUpdatesFilter.expectFlow(bot) {
@@ -40,7 +41,7 @@ internal suspend inline fun <reified T : EncryptedPassportElement> BehaviourCont
 
 suspend fun BehaviourContext.onPassportMessage(
     includeFilterByChatInBehaviourSubContext: Boolean = true,
-    additionalFilter: (suspend (PassportMessage) -> Boolean)? = null,
+    additionalFilter: SimpleFilter<PassportMessage>? = null,
     markerFactory: MarkerFactory<in PassportMessage, Any> = ByChatMessageMarkerFactory,
     scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, PassportMessage>
 ) = onPassportMessageWith<EncryptedPassportElement>(